+++ experimental/packages/opentelemetry-exporter-metrics-otlp-http/test/node/CollectorMetricExporter.test.ts
@@ -173,7 +173,9 @@ describe('OTLPMetricExporter - node with json over http', () => {
       histogram.record(7);
       histogram.record(14);
 
-      metrics = await collect();
+      const { resourceMetrics, errors } = await collect();
+      assert.strictEqual(errors.length, 0);
+      metrics = resourceMetrics;
     });
 
     it('should open the connection', done => {